Exclude files other than "*.cgi" at the top level

+++ mt/mt-watcher.pl
@@ -3,6 +3,7 @@
 use strict;
 use warnings;
 
+use Cwd qw(getcwd);
 use Filesys::Notify::Simple;
 
 if ( $ENV{DISABLE_MT_WATCHER} ) {
@@ -13,13 +14,23 @@ if ( $ENV{DISABLE_MT_WATCHER} ) {
 # wait for `make me` to complete
 sleep(5);
 
-my @files   = ( qw(addons extlib lib plugins), glob('*.cgi') );
+my $mt_home = $ENV{MT_HOME} || getcwd();
+$mt_home =~ s{/+$}{};    # remove trailing slash
+
+my @files   = ( map( {"$mt_home/$_"} qw(addons extlib lib plugins) ), glob("$mt_home/*.cgi") );
 my $watcher = Filesys::Notify::Simple->new( \@files );
 while (1) {
     $watcher->wait(
         sub {
-            my @events = @_;
-            print( join( "\n", map { $_->{path} } @events ) . "\n" );
+            my @paths = grep {
+
+                # exclude files other than "*.cgi" at the top level
+                $_ =~ m{$mt_home/(?:[^/]+\.cgi|.*?/)};
+            } map { $_->{path} } @_;
+
+            return unless @paths;
+
+            print( join( "\n", @paths ) . "\n" );
             system('docker kill -s HUP mt_mt_1');
 
             # throttling
